requirements.txt
Django==2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1
pytz==2019.1
